Fix frontend routes
Api middleware wrapping had broken the frontend routing. Putting the
front end routes first now to avoid having to use wrap-routes to avoid
the api routes matching frontend requests (read somewhere that it's
slightly faster this way, not that it matters much in this application).

Had to remove the "Page not found" route so that it doesn't block the
api routes now.